+++ Bloch/Artifacts/UnlockedPotentialArtifact.cs
@@ -1,6 +1,8 @@
-﻿using Nanoray.PluginManager;
+﻿using HarmonyLib;
+using Nanoray.PluginManager;
 using Nickel;
 using System.Collections.Generic;
+using System.Linq;
 using System.Reflection;
 
 namespace Shockah.Bloch;
@@ -21,8 +23,30 @@ internal sealed class UnlockedPotentialArtifact : Artifact, IRegisterable
 			Name = ModEntry.Instance.AnyLocalizations.Bind(["artifact", "UnlockedPotential", "name"]).Localize,
 			Description = ModEntry.Instance.AnyLocalizations.Bind(["artifact", "UnlockedPotential", "description"]).Localize
 		});
+		
+		ModEntry.Instance.Harmony.Patch(
+			original: AccessTools.AllAssemblies()
+				.First(a => (a.GetName().Name ?? a.GetName().FullName) == "Kokoro")
+				.GetType("Shockah.Kokoro.SpontaneousManager")!
+				.GetNestedType("TriggerAction", AccessTools.all)!
+				.GetMethod("Begin", AccessTools.all)!,
+			postfix: new HarmonyMethod(MethodBase.GetCurrentMethod()!.DeclaringType!, nameof(SpontaneousManager_TriggerAction_Begin_Postfix))
+		);
 	}
 
 	public override List<Tooltip>? GetExtraTooltips()
-		=> new SpontaneousManager.TriggerAction { Action = new ADummyAction() }.GetTooltips(DB.fakeState);
+		=> ModEntry.Instance.KokoroApi.Actions.MakeSpontaneousAction(new ADummyAction()).GetTooltips(DB.fakeState);
+
+	private static void SpontaneousManager_TriggerAction_Begin_Postfix(CardAction __instance, G g, State s, Combat c)
+	{
+		if (s.EnumerateAllArtifacts().FirstOrDefault(a => a is MuscleMemoryArtifact) is not { } artifact)
+			return;
+
+		c.QueueImmediate(ModEntry.Instance.KokoroApi.Actions.GetWrappedCardActions(__instance).Select(action =>
+		{
+			if (string.IsNullOrEmpty(action.artifactPulse))
+				action.artifactPulse = artifact.Key();
+			return action;
+		}));
+	}
 }