+++ Natasha/ExternalAPIs/IBlochApi.cs
@@ -1,7 +1,0 @@
-﻿namespace Shockah.Natasha;
-
-public interface IBlochApi
-{
-	CardAction MakeOnTurnEndAction(CardAction action);
-	CardAction MakeSpontaneousAction(CardAction action);
-}