+++ Natasha/ExternalAPIs/Kokoro/Actions.cs
@@ -31,6 +31,9 @@ public partial interface IKokoroApi
 
 		void RegisterWrappedActionHook(IWrappedActionHook hook, double priority);
 		void UnregisterWrappedActionHook(IWrappedActionHook hook);
+		
+		CardAction MakeSpontaneousAction(CardAction action);
+		CardAction MakeOnTurnEndAction(CardAction action);
 	}
 }