+++ Natasha/ModEntry.cs
@@ -17,7 +17,6 @@ public sealed class ModEntry : SimpleMod
 	internal readonly IHarmony Harmony;
 	internal readonly HookManager<INatashaHook> HookManager;
 	internal readonly IKokoroApi KokoroApi;
-	internal readonly IBlochApi BlochApi;
 	internal readonly ILocalizationProvider<IReadOnlyList<string>> AnyLocalizations;
 	internal readonly ILocaleBoundNonNullLocalizationProvider<IReadOnlyList<string>> Localizations;
 
@@ -104,7 +103,6 @@ public sealed class ModEntry : SimpleMod
 		Harmony = helper.Utilities.Harmony;
 		HookManager = new();
 		KokoroApi = helper.ModRegistry.GetApi<IKokoroApi>("Shockah.Kokoro")!;
-		BlochApi = helper.ModRegistry.GetApi<IBlochApi>("Shockah.Bloch")!;
 
 		this.AnyLocalizations = new JsonLocalizationProvider(
 			tokenExtractor: new SimpleLocalizationTokenExtractor(),